+++ src/styles/deployer.less
@@ -3,6 +3,8 @@
 @import "palette.less";
 @import "common.less";
 @import "fonts.less";
+
+// Reusable components
 @import "components/wizard.less";
 @import "components/installintro.less";
 @import "components/modelpicker.less";
@@ -21,5 +23,7 @@
 @import "components/connetioncredsinfo.less";
 @import "components/messages.less";
 @import "components/loadingmask.less";
-@import "components/cloudsettings.less";
 @import "components/viewserverdetails";
+
+// Pages
+@import "pages/ServerRoleSummary.less";
